主要是id 类型改为 int

+++ app/src/main/java/com/hefengbao/jingmo/data/database/dao/ChineseWisecrackDao.kt
@@ -13,20 +13,20 @@ interface ChineseWisecrackDao {
     suspend fun insert(entity: ChineseWisecrackEntity)
 
     @Query("select * from chinese_wisecracks where id = :id")
-    suspend fun getChineseWisecrack(id: Long): ChineseWisecrackEntity
+    suspend fun getChineseWisecrack(id: Int): ChineseWisecrackEntity
 
     @Query("select id from chinese_wisecracks where id > :id order by id asc limit 1")
-    suspend fun getNextId(id: Long): Long
+    suspend fun getNextId(id: Int): Int
 
     @Query("select id from chinese_wisecracks where id < :id order by id desc limit 1")
-    suspend fun getPrevId(id: Long): Long
+    suspend fun getPrevId(id: Int): Int
 
     @Query("select * from chinese_wisecracks where riddle like :query or answer like :query")
     fun searchWisecrackList(query: String): Flow<List<ChineseWisecrackEntity>>
 
     @Query("select id from chinese_wisecracks where id > :id and (riddle like :query or answer like :query) order by id asc limit 1")
-    suspend fun getSearchNextId(id: Long, query: String): Long
+    suspend fun getSearchNextId(id: Int, query: String): Int
 
     @Query("select id from chinese_wisecracks where id < :id and (riddle like :query or answer like :query) order by id desc limit 1")
-    suspend fun getSearchPrevId(id: Long, query: String): Long
+    suspend fun getSearchPrevId(id: Int, query: String): Int
 }
