+++ app/src/main/java/com/hefengbao/jingmo/data/database/dao/PoemSentenceDao.kt
@@ -16,23 +16,23 @@ interface PoemSentenceDao {
 
     @Transaction
     @Query("select * from poem_sentences where id = :id")
-    suspend fun getSentenceWithPoem(id: Long): SentenceWithPoem
+    suspend fun getSentenceWithPoem(id: Int): SentenceWithPoem
 
     @Query("select * from poem_sentences where id = :id")
-    fun getSentence(id: Long): Flow<PoemSentenceEntity>
+    fun getSentence(id: Int): Flow<PoemSentenceEntity>
 
     @Query("select id from poem_sentences where id > :id order by id asc limit 1")
-    suspend fun getNextId(id: Long): Long
+    suspend fun getNextId(id: Int): Int
 
     @Query("select id from poem_sentences where id < :id order by id desc limit 1")
-    suspend fun getPrevId(id: Long): Long
+    suspend fun getPrevId(id: Int): Int
 
     @Query("select * from poem_sentences where content like :query")
     fun searchSentencesList(query: String): Flow<List<PoemSentenceEntity>>
 
     @Query("select id from poem_sentences where id > :id and content like :query order by id asc limit 1")
-    suspend fun getSearchNextId(id: Long, query: String): Long
+    suspend fun getSearchNextId(id: Int, query: String): Int
 
     @Query("select id from poem_sentences where id < :id and content like :query order by id desc limit 1")
-    suspend fun getSearchPrevId(id: Long, query: String): Long
+    suspend fun getSearchPrevId(id: Int, query: String): Int
 }
