+++ app/src/main/java/com/hefengbao/jingmo/data/repository/PoemSentenceRepository.kt
@@ -5,11 +5,11 @@ import com.hefengbao.jingmo.data.database.entity.SentenceWithPoem
 import kotlinx.coroutines.flow.Flow
 
 interface PoemSentenceRepository {
-    suspend fun getSentenceWithPoem(id: Long): SentenceWithPoem
-    suspend fun getNextId(id: Long): Long
-    suspend fun getPrevId(id: Long): Long
+    suspend fun getSentenceWithPoem(id: Int): SentenceWithPoem
+    suspend fun getNextId(id: Int): Int
+    suspend fun getPrevId(id: Int): Int
     fun searchSentencesList(query: String): Flow<List<PoemSentenceEntity>>
-    suspend fun getSentence(id: Long): Flow<PoemSentenceEntity>
-    suspend fun getSearchNextId(id: Long, query: String): Long
-    suspend fun getSearchPrevId(id: Long, query: String): Long
+    suspend fun getSentence(id: Int): Flow<PoemSentenceEntity>
+    suspend fun getSearchNextId(id: Int, query: String): Int
+    suspend fun getSearchPrevId(id: Int, query: String): Int
 }
