+++ app/src/main/java/com/hefengbao/jingmo/ui/screen/chinesewisecrack/nav/ChineseWisecrackNav.kt
@@ -15,8 +15,8 @@ fun NavController.navigateToChineseWisecrackGraph() {
 
 fun NavGraphBuilder.chineseWisecrackGraph(
     onBackClick: () -> Unit,
-    onCaptureClick: (Long) -> Unit,
-    onSearchItemClick: (Long, String) -> Unit,
+    onCaptureClick: (Int) -> Unit,
+    onSearchItemClick: (Int, String) -> Unit,
     nestGraph: NavGraphBuilder.() -> Unit
 ) {
     navigation(
