+++ app/src/main/java/com/hefengbao/jingmo/ui/screen/idiom/IdiomSearchShowScreen.kt
@@ -34,7 +34,7 @@ import com.hefengbao.jingmo.ui.component.SimpleScaffold
 @Composable
 fun IdiomSearchShowRoute(
     onBackClick: () -> Unit,
-    onCaptureClick: (Long) -> Unit,
+    onCaptureClick: (Int) -> Unit,
     viewModel: IdiomSearchShowViewModel = hiltViewModel()
 ) {
 
@@ -74,10 +74,10 @@ fun IdiomSearchShowRoute(
 private fun IdiomSearchShowScreen(
     modifier: Modifier = Modifier,
     onBackClick: () -> Unit,
-    onCaptureClick: (Long) -> Unit,
+    onCaptureClick: (Int) -> Unit,
     idiom: IdiomEntity?,
-    prevId: Long?,
-    nextId: Long?,
+    prevId: Int?,
+    nextId: Int?,
     onPrevClick: () -> Unit,
     onNextClick: () -> Unit,
     query: String
@@ -188,14 +188,14 @@ private fun IdiomSearchShowScreen(
                 ) {
                     IconButton(
                         onClick = onPrevClick,
-                        enabled = prevId != 0L
+                        enabled = prevId != 0
                     ) {
                         Icon(imageVector = Icons.Default.ArrowBack, contentDescription = null)
                     }
 
                     IconButton(
                         onClick = onNextClick,
-                        enabled = nextId != 0L
+                        enabled = nextId != 0
                     ) {
                         Icon(imageVector = Icons.Default.ArrowForward, contentDescription = null)
                     }
