+++ app/src/main/java/com/hefengbao/jingmo/ui/screen/idiom/nav/IdiomListNav.kt
@@ -15,8 +15,8 @@ fun NavController.navigateToIdiomListGraph() {
 
 fun NavGraphBuilder.idiomListGraph(
     onBackClick: () -> Unit,
-    onItemClick: (Long) -> Unit,
-    onSearchItemClick: (Long, String) -> Unit,
+    onItemClick: (Int) -> Unit,
+    onSearchItemClick: (Int, String) -> Unit,
     nestGraph: NavGraphBuilder.() -> Unit
 ) {
     navigation(
